+++ src/util/tls.ts
@@ -52,9 +52,11 @@ async function _readKeyAndCert(keyPath: string, certPath: string):
       }
 
       try {
-        const [cert, key] =
+        const results =
             (await Promise.all([fs.readFile(certPath), fs.readFile(keyPath)]))
                 .map(buffer => buffer.toString().trim());
+        const cert = results[0];
+        const key = results[1];
         if (key && cert) {
           return {cert, key};
         }